Rewrite repo in typescript (#72)
* Adding typescript for Viro3DObject to test out typescript conversion.

* For some reason the dist/components directory doesn't get picked up when using the git repo as a target on the typescript branch, attempting to work out why.

* Adding dist to the explicit files list in package.json

* Moving output from tsc to be next to JS files.

* Trying different tsconfig settings

* chore: update .gitignore

* fix: typescript file reverted to JS

* feat: convert entire repo to TypeScript

In order to enhance the developer experience for both the ViroCommunity
maintainers and also the end developer, I've converted the repo to
typescript.

There were many instances where common event handlers were copy/pasted
across Viro Components. This makes the repo hard to maintain.

The end user can also get types, commends, and docs while developing
using an IDE. This also will help the end user avoid usage mistakes,
which in turn should reduce the load on the maintainers.

* feat: export from index.ts correctly

* fix: fix typescript definitions not building correctly

* fix: fix some issues with migrated components

* fix: issues with components from TS migration

* Update tsconfig for easier development

Adds an output directory to the tsconfig so the source code is separate
from the generated js and d.ts files. This will help with ease of
development

* Add Resources folder to dist

Without the Resources folder being copied by tsc, this would not
resolve the dependencies correctly, since there are a few default
placeholder images.

* Fix typescript build issue with npm

Co-authored-by: Paul McCabe <phmccabe@googlemail.com>